Traspasos (Completo)

+++ EC-Admin/EC-Admin/Forms/Promocion/frmNuevaPromocion.cs
@@ -22,7 +22,7 @@ namespace EC_Admin.Forms
             chbExistencias.Checked = true;
         }
 
-        public void AsignarProducto(int id, string nombre, string marca, int cant)
+        public void AsignarProducto(int id, string nombre, string marca, decimal cant)
         {
             this.id = id;
             this.cant = cant;
